Fix en funcionalidad de lightbox luego de regenerar el listado de mensajes por Ajax.
Agregado pasaje de filtros a acciones de marcar como leido, etc...

Correjido problema al presionar eliminar sin ningun mensaje seleccionado.

+++ WEB-INF/classes/modules/common/classes/InternalMailPeer.php
@@ -108,7 +108,10 @@ class InternalMailPeer extends BaseInternalMailPeer {
 	}
 	
 	public static function delete($ids) {
-		return InternalMailQuery::create()->filterByPrimaryKeys($ids)->delete();
+		if (!empty($ids))
+			return InternalMailQuery::create()->filterByPrimaryKeys($ids)->delete();
+		else
+			return 0;
 	}
 	
 	public static function markAsRead($ids) {
